requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-chroma==0.1.4
chromadb==0.5.3
pymupdf==1.24.5
python-dotenv==1.0.1
pydantic==2.7.4
sentence-transformers==3.0.1
google-genai==0.8.0